backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
dj-rest-auth==7.0.1
Django==5.1.5
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.3.1
django-extensions==4.1
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
filelock==3.14.0
gunicorn==23.0.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.2
packaging==25.0
platformdirs==4.2.1
progressbar2==4.5.0
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-utils==3.9.1
pytz==2025.2
requests==2.31.0
setuptools==69.5.1
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.14.1
tzdata==2024.1
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1